+++ packages/elements/stories/Line.stories.js
@@ -6,6 +6,8 @@ import Line03Story from "./Line/Line03.svelte";
 import Line04Story from "./Line/Line04.svelte";
 import Line05Story from "./Line/Line05.svelte";
 
+import TrumpApprovalStory from "./Line/TrumpApproval.svelte";
+
 export default {
   title: "Line",
   decorators: [withKnobs],
@@ -54,4 +56,8 @@ export const Line05 = () => ({
     supertitle: "Over Time",
     description: "The lines go up and down like a mountain!",
   },
+});
+
+export const ApprovalPollModelExample = () => ({
+  Component: TrumpApprovalStory,
 });